[garnet][rust] Make zx_time_t and zx_duration_t signed types
This change is part of a hard transition.

Zircon's C time types, zx_time_t and zx_duration_t, are now signed.

Update corresponding Rust types to be signed as well.

Change Rust's ZX_TIME_INFINITE from ::std::u64::MAX to ::std::i64::MAX
to match Zircon's ZX_TIME_INFINITE.

Change several time constants from unsigned to signed. For others,
cast when constructing duration/time objects.

Remove from() function that converts Duration to stdtime::Duration as
it's only used time.rs tests and there isn't a single conversion that
makes sense for all uses now that Durations are signed.

Test: CQ; built and ran topaz system on Eve

ZX-2100 #comment

+++ public/rust/crates/fuchsia-zircon/src/lib.rs
@@ -231,7 +231,7 @@ mod tests {
         let ticks2 = ticks_get();
 
         // The number of ticks should have increased by at least 1 ms worth
-        let sleep_ticks = sleep_time.millis() * ticks_per_second() / 1000;
+        let sleep_ticks = (sleep_time.millis() as u64) * ticks_per_second() / 1000;
         assert!(ticks2 >= (ticks1 + sleep_ticks));
     }
